Adjustments to make the font work
