+++ src/components/UI/Form/Textarea.styled.js
@@ -1,11 +1,13 @@
 import styled from 'styled-components';
 
 const Textarea = styled.textarea`
+	font-family: Arimo, sans-serif;
+	font-size: 1.6rem;
 	width: 100%;
 	margin-top: 5px;
 	padding: 5px 5px;
 	height: 10rem;
-	font-family: Arial, Helvetica, sans-serif;
+
 	resize: none;
 `;
